refactor(ocean): extract palettes, shaders, textures and helpers
First slice of breaking up the 1800-line Ocean3D.tsx: the leaf pieces
with no scene state move out verbatim - day/night palettes, the sky
and water GLSL, canvas texture painters (glow, moon), seeded RNG and
the rock jitter helpers. Scene builders are next, behind handle-object
returns, so this commit is import shuffling only.
