+++ components/Ocean3D/Ocean3D.tsx
@@ -20,8 +20,6 @@ import {
 	Fog,
 	Group,
 	HemisphereLight,
-	IcosahedronGeometry,
-	Material,
 	Mesh,
 	MeshBasicMaterial,
 	MeshStandardMaterial,
@@ -32,7 +30,6 @@ import {
 	PointLight,
 	Points,
 	PointsMaterial,
-	SRGBColorSpace,
 	Scene,
 	ShaderMaterial,
 	SphereGeometry,
@@ -45,6 +42,11 @@ import {
 } from "three";
 import { useTheme } from "@/providers/ThemeProvider";
 import { getMoonPhase, getIlluminationFraction } from "@/lib/moonPhase";
+import { DAY, NIGHT } from "./palette";
+import { hashString, mulberry32 } from "./random";
+import { jitterByPosition, makeRock } from "./geometry";
+import { makeGlowTexture, makeMoonTexture } from "./textures";
+import { SKY_VERT, SKY_FRAG, WATER_VERT, WATER_FRAG } from "./shaders";
 import styles from "./Ocean3D.module.css";
 
 // Feature flag: render the actual lunar phase instead of a permanent full
@@ -64,386 +66,6 @@ const SEABED_Y = -18;
 const SEABED_MIN_Y = -7;
 const CAMERA_FLOOR_CLEARANCE = 2.5;
 const SUN_X = -70;
-
-interface Palette {
-	skyTop: number;
-	skyHorizon: number;
-	fog: number;
-	waterNear: number;
-	waterFar: number;
-	glassShallow: number;
-	glassDeep: number;
-	foam: number;
-	sun: number;
-	sunGlow: number;
-	glowOpacity: number;
-	rock: number;
-	underGeo: number;
-	hemiSky: number;
-	hemiGround: number;
-	hemiIntensity: number;
-	dirColor: number;
-	dirIntensity: number;
-	starsOpacity: number;
-	beaconIntensity: number;
-	glint: number;
-	seaweed: number;
-	starfish: number;
-}
-
-const DAY: Palette = {
-	skyTop: 0x837093,
-	skyHorizon: 0xf2dcc4,
-	fog: 0x5ea7d8,
-	waterNear: 0x4f9dd6,
-	waterFar: 0xa8cbdd,
-	glassShallow: 0x5fb0e4,
-	glassDeep: 0x123f6b,
-	foam: 0xeaf6ff,
-	sun: 0xfff0d0,
-	sunGlow: 0xffd9a0,
-	glowOpacity: 0.6,
-	rock: 0x97a98f,
-	underGeo: 0xbcdcee,
-	hemiSky: 0xe2d8c8,
-	hemiGround: 0x4a7896,
-	hemiIntensity: 1.1,
-	dirColor: 0xffe4be,
-	dirIntensity: 1.4,
-	starsOpacity: 0,
-	beaconIntensity: 2,
-	glint: 0.55,
-	seaweed: 0x3a7a5e,
-	starfish: 0xcf7d52,
-};
-
-const NIGHT: Palette = {
-	skyTop: 0x081426,
-	skyHorizon: 0x1c3a54,
-	fog: 0x0e2638,
-	waterNear: 0x1c4470,
-	waterFar: 0x1f4060,
-	glassShallow: 0x16406a,
-	glassDeep: 0x081826,
-	foam: 0x7396ac,
-	sun: 0xf2f4e0,
-	sunGlow: 0xbcd4e8,
-	glowOpacity: 0.3,
-	rock: 0x5a6a60,
-	underGeo: 0x3a5e7a,
-	hemiSky: 0x4f739c,
-	hemiGround: 0x1c3452,
-	hemiIntensity: 1.65,
-	dirColor: 0xc4d8ec,
-	dirIntensity: 1.25,
-	starsOpacity: 0.9,
-	beaconIntensity: 30,
-	glint: 0.3,
-	seaweed: 0x1f4a42,
-	starfish: 0x70453a,
-};
-
-// FNV-1a: turns the page pathname into a stable scatter seed, so every URL
-// gets its own arrangement of stones, fish and seabed — same every visit.
-function hashString(str: string): number {
-	let h = 2166136261;
-	for (let i = 0; i < str.length; i++) {
-		h ^= str.charCodeAt(i);
-		h = Math.imul(h, 16777619);
-	}
-	return h >>> 0;
-}
-
-function mulberry32(seed: number) {
-	let a = seed;
-	return () => {
-		a |= 0;
-		a = (a + 0x6d2b79f5) | 0;
-		let t = Math.imul(a ^ (a >>> 15), 1 | a);
-		t = (t + Math.imul(t ^ (t >>> 7), 61 | t)) ^ t;
-		return ((t ^ (t >>> 14)) >>> 0) / 4294967296;
-	};
-}
-
-// Position-hashed jitter: duplicated vertices (non-indexed polyhedra) get
-// identical offsets, so faces stay welded while the silhouette gets craggy.
-function jitterByPosition(geometry: BufferGeometry, amount: number) {
-	const pos = geometry.getAttribute("position");
-	for (let i = 0; i < pos.count; i++) {
-		const x = pos.getX(i);
-		const y = pos.getY(i);
-		const z = pos.getZ(i);
-		const h = (mx: number, my: number, mz: number) => {
-			const s = Math.sin(x * mx + y * my + z * mz) * 43758.5453;
-			return (s - Math.floor(s)) - 0.5;
-		};
-		pos.setXYZ(
-			i,
-			x + h(127.1, 311.7, 74.7) * amount,
-			y + h(269.5, 183.3, 246.1) * amount,
-			z + h(113.5, 271.9, 124.6) * amount,
-		);
-	}
-	pos.needsUpdate = true;
-}
-
-function makeRock(material: Material, jitter = 0.35): Mesh {
-	const geometry = new IcosahedronGeometry(1, 1);
-	jitterByPosition(geometry, jitter);
-	return new Mesh(geometry, material);
-}
-
-function makeGlowTexture(): CanvasTexture {
-	const size = 256;
-	const canvas = document.createElement("canvas");
-	canvas.width = size;
-	canvas.height = size;
-	const ctx = canvas.getContext("2d");
-	if (ctx) {
-		const g = ctx.createRadialGradient(size / 2, size / 2, 0, size / 2, size / 2, size / 2);
-		g.addColorStop(0, "rgba(255,255,255,0.9)");
-		g.addColorStop(0.3, "rgba(255,255,255,0.35)");
-		g.addColorStop(1, "rgba(255,255,255,0)");
-		ctx.fillStyle = g;
-		ctx.fillRect(0, 0, size, size);
-	}
-	return new CanvasTexture(canvas);
-}
-
-// Moon with the old site's craters, optionally shaded by the real lunar
-// phase (half-circle + ellipse terminator, clipped to the disc).
-function makeMoonTexture(phase: number | null): CanvasTexture {
-	const size = 256;
-	const c = size / 2;
-	const r = size * 0.48;
-	const canvas = document.createElement("canvas");
-	canvas.width = size;
-	canvas.height = size;
-	const ctx = canvas.getContext("2d");
-	if (ctx) {
-		ctx.fillStyle = "#f4f1de";
-		ctx.beginPath();
-		ctx.arc(c, c, r, 0, Math.PI * 2);
-		ctx.fill();
-
-		// Soft-edged maria laid out like the real near side: Procellarum on
-		// the left limb, Imbrium upper left, Serenitatis/Tranquillitatis in
-		// the upper middle, Fecunditatis/Nectaris lower right.
-		ctx.save();
-		ctx.beginPath();
-		ctx.arc(c, c, r, 0, Math.PI * 2);
-		ctx.clip();
-		const mare = (mx: number, my: number, mr: number, alpha: number) => {
-			const px = c + mx * r;
-			const py = c + my * r;
-			const g = ctx.createRadialGradient(px, py, 0, px, py, mr * r);
-			g.addColorStop(0, `rgba(166, 163, 145, ${alpha})`);
-			g.addColorStop(0.7, `rgba(166, 163, 145, ${alpha * 0.8})`);
-			g.addColorStop(1, "rgba(166, 163, 145, 0)");
-			ctx.fillStyle = g;
-			ctx.beginPath();
-			ctx.arc(px, py, mr * r, 0, Math.PI * 2);
-			ctx.fill();
-		};
-		mare(-0.52, -0.12, 0.3, 0.5); // Oceanus Procellarum (upper)
-		mare(-0.48, 0.22, 0.24, 0.45); // Oceanus Procellarum (lower)
-		mare(-0.22, -0.38, 0.28, 0.55); // Mare Imbrium
-		mare(0.16, -0.36, 0.2, 0.5); // Mare Serenitatis
-		mare(0.34, -0.1, 0.21, 0.5); // Mare Tranquillitatis
-		mare(0.52, 0.14, 0.15, 0.45); // Mare Fecunditatis
-		mare(0.28, 0.28, 0.11, 0.4); // Mare Nectaris
-		mare(-0.18, 0.32, 0.13, 0.35); // Mare Nubium
-		// A few small craters in the bright highlands.
-		ctx.fillStyle = "rgba(160, 157, 140, 0.4)";
-		const craters: [number, number, number][] = [
-			[-0.3, 0.05, 0.045], // Copernicus
-			[0.05, 0.5, 0.035], // Albategnius-ish
-			[-0.55, 0.5, 0.03],
-			[0.55, -0.42, 0.03],
-		];
-		for (const [cx, cy, cr] of craters) {
-			ctx.beginPath();
-			ctx.arc(c + cx * r, c + cy * r, cr * r, 0, Math.PI * 2);
-			ctx.fill();
-		}
-		// Tycho: bright spot near the southern limb.
-		const tycho = ctx.createRadialGradient(
-			c - 0.08 * r,
-			c + 0.62 * r,
-			0,
-			c - 0.08 * r,
-			c + 0.62 * r,
-			0.12 * r,
-		);
-		tycho.addColorStop(0, "rgba(255, 253, 240, 0.9)");
-		tycho.addColorStop(1, "rgba(255, 253, 240, 0)");
-		ctx.fillStyle = tycho;
-		ctx.beginPath();
-		ctx.arc(c - 0.08 * r, c + 0.62 * r, 0.12 * r, 0, Math.PI * 2);
-		ctx.fill();
-		ctx.restore();
-
-		if (phase !== null && getIlluminationFraction(phase) < 0.99) {
-			ctx.save();
-			ctx.beginPath();
-			ctx.arc(c, c, r, 0, Math.PI * 2);
-			ctx.clip();
-			const terminatorRx = r * Math.abs(Math.cos(phase * 2 * Math.PI));
-			ctx.fillStyle = "rgba(10, 22, 40, 0.85)";
-			ctx.beginPath();
-			if (phase < 0.5) {
-				ctx.arc(c, c, r, Math.PI * 0.5, Math.PI * 1.5);
-				ctx.ellipse(c, c, terminatorRx, r, 0, Math.PI * 1.5, Math.PI * 0.5);
-			} else {
-				ctx.arc(c, c, r, Math.PI * 1.5, Math.PI * 0.5);
-				ctx.ellipse(c, c, terminatorRx, r, 0, Math.PI * 0.5, Math.PI * 1.5);
-			}
-			ctx.closePath();
-			ctx.fill();
-			ctx.restore();
-		}
-	}
-	const texture = new CanvasTexture(canvas);
-	texture.colorSpace = SRGBColorSpace;
-	return texture;
-}
-
-const SKY_VERT = /* glsl */ `
-	varying vec3 vWorldPos;
-	void main() {
-		vWorldPos = (modelMatrix * vec4(position, 1.0)).xyz;
-		gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
-	}
-`;
-
-const SKY_FRAG = /* glsl */ `
-	uniform vec3 uTop;
-	uniform vec3 uHorizon;
-	varying vec3 vWorldPos;
-	void main() {
-		float h = normalize(vWorldPos).y;
-		vec3 col = mix(uHorizon, uTop, smoothstep(-0.02, 0.5, h));
-		gl_FragColor = vec4(col, 1.0);
-	}
-`;
-
-const WATER_VERT = /* glsl */ `
-	uniform float uTime;
-	varying vec3 vPos;
-	varying float vWave;
-	void main() {
-		vec3 p = position;
-		float w = sin(p.x * 0.045 + uTime * 0.6) * 0.42
-			+ sin(p.x * 0.11 - uTime * 0.42 + p.z * 0.06) * 0.24
-			+ sin(p.z * 0.16 + uTime * 0.5 + p.x * 0.02) * 0.2;
-		// Only ~9 world units of waterline are visible at the cut plane, so
-		// long swells read as a straight line there: they fade out toward the
-		// cut while a fine ripple fades in.
-		float ripple = sin(p.x * 0.8 + uTime * 0.9) * 0.12
-			+ sin(p.x * 0.45 - uTime * 0.55 + p.z * 0.8) * 0.1;
-		// Only the top sheet undulates; the wall's deep bottom row (the sole
-		// vertices below y=0) stays put. The wall's top row IS the sheet's
-		// edge row, so the waterline is welded shut by construction.
-		if (p.y > -1.0) {
-			p.y += w * smoothstep(30.0, 8.0, p.z) + ripple * smoothstep(8.0, 29.9, p.z);
-		}
-		vWave = w * 0.7 + ripple;
-		vPos = p;
-		gl_Position = projectionMatrix * modelViewMatrix * vec4(p, 1.0);
-	}
-`;
-
-const WATER_FRAG = /* glsl */ `
-	uniform vec3 uNear;
-	uniform vec3 uFar;
-	uniform vec3 uHorizon;
-	uniform vec3 uSunColor;
-	uniform vec3 uShallow;
-	uniform vec3 uDeep;
-	uniform vec3 uFoam;
-	uniform float uSunX;
-	uniform float uTime;
-	uniform float uGlint;
-	uniform vec2 uBeamDir;
-	uniform float uBeamStrength;
-	varying vec3 vPos;
-	varying float vWave;
-	void main() {
-		// Per-facet normals from screen-space derivatives: chunky low-poly
-		// shading, and a reliable way to tell the vertical cross-section
-		// wall apart from the top sheet.
-		vec3 facetNormal = normalize(cross(dFdx(vPos), dFdy(vPos)));
-
-		if (abs(facetNormal.y) < 0.35) {
-			// Cross-section wall: the "aquarium glass" front of the block.
-			float surf = sin(vPos.x * 0.8 + uTime * 0.9) * 0.12
-				+ sin(vPos.x * 0.45 - uTime * 0.55 + 24.0) * 0.1;
-			float depth = clamp(-vPos.y / 26.0, 0.0, 1.0);
-			vec3 wallCol = mix(uShallow, uDeep, smoothstep(0.0, 1.0, depth));
-			float alpha = mix(0.5, 0.68, depth);
-			float line = smoothstep(0.25, 0.0, abs(vPos.y - surf));
-			wallCol = mix(wallCol, uFoam, line * 0.85);
-			alpha = max(alpha, line * 0.95);
-			gl_FragColor = vec4(wallCol, alpha);
-			return;
-		}
-
-		float dist = clamp((36.0 - vPos.z) / 220.0, 0.0, 1.0);
-		vec3 col = mix(uNear, uFar, smoothstep(0.0, 0.7, dist));
-		col += vWave * 0.06;
-		float facetLight = clamp(dot(facetNormal, normalize(vec3(-0.35, 0.9, 0.25))), 0.0, 1.0);
-		col *= 0.88 + 0.3 * facetLight;
-		// Grazing-angle fresnel, weighted by distance: near water keeps its
-		// blue even when the camera sits at the waterline; only the far
-		// surface melts into the horizon so there is no hard sky edge.
-		vec3 viewDir = normalize(cameraPosition - vPos);
-		float grazing = pow(1.0 - clamp(abs(viewDir.y), 0.0, 1.0), 4.0);
-		col = mix(col, uHorizon, grazing * (0.2 + 0.6 * smoothstep(0.1, 0.9, dist)));
-		// Foam rings lapping the island shores; footprints match the rock
-		// placements (x, z, radiusX, radiusZ). Living in the shader means
-		// the foam rides the waves instead of floating as a separate mesh.
-		vec4 islands[4];
-		islands[0] = vec4(-55.0, -120.0, 17.5, 12.5);
-		islands[1] = vec4(-42.0, -112.0, 8.5, 7.5);
-		islands[2] = vec4(15.0, -75.0, 10.5, 8.5);
-		islands[3] = vec4(24.0, -71.0, 5.5, 5.5);
-		float foamRing = 0.0;
-		for (int i = 0; i < 4; i++) {
-			vec2 rel = vec2(
-				(vPos.x - islands[i].x) / islands[i].z,
-				(vPos.z - islands[i].y) / islands[i].w
-			);
-			float d = length(rel);
-			float ring = smoothstep(0.8, 1.0, d) * smoothstep(1.3, 1.05, d);
-			foamRing = max(foamRing, ring);
-		}
-		foamRing *= 0.55 + 0.45 * sin(vPos.x * 1.7 + vPos.z * 1.3 - uTime * 1.6);
-		col = mix(col, uFoam, foamRing * 0.75);
-		float cx = uSunX * dist;
-		float streak = exp(-pow(vPos.x - cx, 2.0) / (3.0 + 360.0 * dist * dist));
-		float shimmer = 0.6 + 0.4 * sin(vPos.x * 2.4 + vPos.z * 3.1 + uTime * 2.2);
-		// The shimmer's world-space stripes alias into banding at range:
-		// fade to a smooth glow toward the horizon.
-		shimmer = mix(shimmer, 0.65, smoothstep(0.25, 0.7, dist));
-		col += uSunColor * streak * shimmer * uGlint * dist;
-		// Lighthouse sweep: a warm light pool travels across the waves where
-		// the rotating beam points (bi-directional, so abs of the alignment).
-		vec2 toFrag = vPos.xz - vec2(12.0, -72.0);
-		float lampDist = length(toFrag);
-		float align = abs(dot(toFrag / max(lampDist, 0.001), uBeamDir));
-		float pool = smoothstep(0.978, 0.999, align) * smoothstep(85.0, 12.0, lampDist);
-		col += vec3(1.0, 0.93, 0.75) * pool * uBeamStrength;
-		float alpha = 0.94;
-		if (!gl_FrontFacing) {
-			// Underwater looking up: crossings near the camera stay translucent
-			// so the islands read through the surface right at the waterline,
-			// while long underwater sightlines go opaque so the above-water
-			// world fades out with depth instead of floating in the blue.
-			alpha = mix(0.6, 1.0, smoothstep(12.0, 55.0, length(cameraPosition - vPos)));
-		}
-		gl_FragColor = vec4(col, alpha);
-	}
-`;
 
 interface SceneApi {
 	setDark: (dark: boolean) => void;
