+++ components/Ocean3D/random.ts
@@ -1,0 +1,21 @@
+// FNV-1a: turns the page pathname into a stable scatter seed, so every URL
+// gets its own arrangement of stones, fish and seabed — same every visit.
+export function hashString(str: string): number {
+	let h = 2166136261;
+	for (let i = 0; i < str.length; i++) {
+		h ^= str.charCodeAt(i);
+		h = Math.imul(h, 16777619);
+	}
+	return h >>> 0;
+}
+
+export function mulberry32(seed: number) {
+	let a = seed;
+	return () => {
+		a |= 0;
+		a = (a + 0x6d2b79f5) | 0;
+		let t = Math.imul(a ^ (a >>> 15), 1 | a);
+		t = (t + Math.imul(t ^ (t >>> 7), 61 | t)) ^ t;
+		return ((t ^ (t >>> 14)) >>> 0) / 4294967296;
+	};
+}
